+++ lib/BlockCypher/Crypto/Signer.php
@@ -5,7 +5,6 @@ namespace BlockCypher\Crypto;
 use BitWasp\Bitcoin\Bitcoin;
 use BitWasp\Bitcoin\Crypto\EcAdapter\EcAdapterInterface;
 use BitWasp\Bitcoin\Crypto\Random\Rfc6979;
-use BitWasp\Bitcoin\Key\PrivateKeyFactory;
 use BitWasp\Bitcoin\Key\PrivateKeyInterface;
 use BitWasp\Buffertools\Buffer;
 
@@ -45,7 +44,7 @@ class Signer
     public static function sign($hexDataToSign, $privateKey)
     {
         if (is_string($privateKey)) {
-            $privateKey = self::importPrivateKey($privateKey);
+            $privateKey = PrivateKeyManipulator::importPrivateKey($privateKey);
         }
 
         // Convert hex data to buffer
@@ -66,65 +65,5 @@ class Signer
         //var_dump($sig->getHex());
 
         return $sig->getHex();
-    }
-
-    /**
-     * @param string $plainPrivateKey
-     * @return PrivateKeyInterface
-     */
-    public static function importPrivateKey($plainPrivateKey)
-    {
-        $privateKey = null;
-        $extraMsg = '';
-
-        // TODO: Code Review. Method to detect private key format.
-
-        if ($privateKey === null) {
-            try {
-                $privateKey = PrivateKeyFactory::fromWif($plainPrivateKey);
-            } catch (\Exception $e) {
-                $extraMsg .= " Error trying to import from Wif: " . $e->getMessage();
-            }
-        }
-
-        if ($privateKey === null) {
-            try {
-                $privateKey = PrivateKeyFactory::fromHex($plainPrivateKey);
-            } catch (\Exception $e) {
-                $extraMsg .= " Error trying to import from Hex: " . $e->getMessage();
-            }
-        }
-
-        if ($privateKey === null) {
-            throw new \InvalidArgumentException("Invalid private key format. " . $extraMsg);
-        }
-
-        return $privateKey;
-    }
-
-    /**
-     * @param string $hexPrivateKey
-     * @return PrivateKeyInterface
-     * @throws \Exception
-     */
-    public static function importPrivateKeyFromHex($hexPrivateKey)
-    {
-        // Import from compressed private key
-        $compressed = true;
-        $privateKey = PrivateKeyFactory::fromHex($hexPrivateKey, $compressed);
-
-        return $privateKey;
-    }
-
-    /**
-     * @param string $wifPrivateKey
-     * @return PrivateKeyInterface
-     * @throws \Exception
-     */
-    public static function importPrivateKeyFromWif($wifPrivateKey)
-    {
-        $privateKey = PrivateKeyFactory::fromWif($wifPrivateKey);
-
-        return $privateKey;
     }
 }