+++ sample/index.php
@@ -954,6 +954,57 @@ if (PHP_SAPI == 'cli') {
                             </div>
                         </div>
                     </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8"><h5>Create from private (with only one method call)</h5>
+                            </div>
+                            <div class="col-md-4">
+                                <a href="microtransaction-api/CreateMicroTXFromPrivate.php"
+                                   class="btn btn-primary pull-left execute">
+                                    Try It
+                                    <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/microtransaction-api/CreateMicroTXFromPrivate.html"
+                                   class="btn btn-default pull-right">Source
+                                    <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8"><h5>Create from pubkey (with only one method call)</h5>
+                            </div>
+                            <div class="col-md-4">
+                                <a href="microtransaction-api/CreateMicroTXFromPubkey.php"
+                                   class="btn btn-primary pull-left execute">
+                                    Try It
+                                    <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/microtransaction-api/CreateMicroTXFromPubkey.html"
+                                   class="btn btn-default pull-right">Source
+                                    <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
+                    <li class="list-group-item">
+                        <div class="row">
+                            <div class="col-md-8"><h5>Create from WIF (with only one method call)</h5>
+                            </div>
+                            <div class="col-md-4">
+                                <a href="microtransaction-api/CreateMicroTXFromWif.php"
+                                   class="btn btn-primary pull-left execute">
+                                    Try It
+                                    <i
+                                        class="fa fa-play-circle-o"></i></a>
+                                <a href="doc/microtransaction-api/CreateMicroTXFromWif.html"
+                                   class="btn btn-default pull-right">Source
+                                    <i
+                                        class="fa fa-file-code-o"></i></a>
+                            </div>
+                        </div>
+                    </li>
                 </ul>
             </div>